requirements.txt
# CREST — Python Dependencies
# Install: pip install -r requirements.txt

# ── API & Web ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.3
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────────
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.30

# ── Task Queue ─────────────────────────────────────────────
celery==5.4.0
redis==5.0.4
confluent-kafka==2.4.0

# ── AI / LLM ──────────────────────────────────────────────
anthropic==0.26.0
langchain==0.2.1
langchain-anthropic==0.1.15
langchain-community==0.2.1
langchain-core==0.2.5
llama-index==0.10.40
llama-index-vector-stores-postgres==0.1.9
openai==1.30.1             # Whisper STT + embeddings fallback

# ── NLP ────────────────────────────────────────────────────
spacy==3.7.4
# Run after install: python -m spacy download en_core_web_sm

# ── Search ─────────────────────────────────────────────────
elasticsearch==8.13.0

# ── Utils ──────────────────────────────────────────────────
numpy==1.26.4
python-dotenv==1.0.1
